# src/agent/.kiota/documents/examplesapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Examples API",
    "x-ai-description": "Search for customers, and lots of other stuff",
    "description": "API for ordering assets",
    "version": "0.0.0"
  },
  "tags": [],
  "paths": {
    "/customers": {
      "get": {
        "operationId": "GetCustomers",
        "description": "List all available customers",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "prefer",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-ai-reasoning-instructions": [
          "Invoke this tool to get a list of customers. Provide a name query parameter to filter the results."
        ]
      },
      "post": {
        "operationId": "CreateCustomer",
        "description": "Create a customer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentExamples.Customer"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentExamples.Customer"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentExamples.Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "AgentExamples.Customer": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AgentExamples.Address"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AgentExamples.Address"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://0l7j96tl-5056.use.devtunnels.ms",
      "description": "Assets API",
      "variables": {}
    }
  ]
}